requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
langchain==0.0.335
huggingface-hub==0.19.2
PyPDF2==3.0.1
sentence-transformers==2.3.1